backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
sympy==1.12
numpy==1.26.2
pydantic==2.5.0
pytesseract==0.3.10
